Upstream version 1.1+svn10616

# gpg: Signature made Sa 02 Nov 2013 20:42:28 CET
# gpg:                using DSA key 0x9FCF2CCD3F3E6426
# gpg: please do a --check-trustdb
# gpg: Good signature from "Guido Günther <agx@sigxcpu.org>"
# gpg:                 aka "Guido Guenther <agx@debian.org>"
# gpg:                 aka "Guido Guenther <agx@sigxcpu.org>"

+++ src/base64.h
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2006-2008 Apple Inc. All rights reserved.
+ * Copyright (c) 2006-2013 Apple Inc. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,5 +14,6 @@
  * limitations under the License.
  **/
 
-char *base64_encode(const unsigned char *value, int vlen);
-unsigned char *base64_decode(const char *value, int *rlen);
+#include <stddef.h>
+char *base64_encode(const unsigned char *value, size_t vlen);
+unsigned char *base64_decode(const char *value, size_t *rlen);
